feat(*): profile and link-rel urls can be independent of the portal (#2021)
The goal of this PR is to resolve #1996 and make profiles and link
relations easier to understand.

The following is changed:
- profile and link-rel URLs no longer need to be prefixed with
`https://api.otto.de/portal/..`. Instead, it is advised to prefix those
URLs with a base path representing the APIs base path.
- profile and link-rel URLs no longer need to be resolvable. However,
like with problems, it is encouraged to provide reasonable URLs.
- Because profiles no longer need to be resolvable, a link to a profile
can no longer be a MUST. Therefore removed all parts where links to
profile were used.
- Additionally removed "[MAY use type and profile attributes in
hyperlinks](https://github.com/otto-ec/ottoapi_guidelines/blob/6d2e43d762eb96282bf890b8bf765f3f82017947/content/guidelines/030_REST-GUIDELINES/020_Hypermedia/040_Profiles/040_may-use-type-and-profile-attributes-in-hyperlinks.md)".
This rule is a contradiction to [MUST document link
cardinality](https://api.otto.de/portal/guidelines/r100063). As all
profiles need to be present in the "profile" link rel, adding a new
version would change the cardinality of the profile link relation. In
addition to this, the profile link must not be resolvable anymore.
- Integrated the "Profile" section into the "Versioning" section because
only one rule has been left in profiles. This also resolves #1401.
- Moved the description of what a profile is from a collapsed section
that is hardly ever read to the rule that enforces the use of profiles.
- Aligned the statement that the rule only is a must rule to the way it
is stated in the rule [MUST use common paging query
parameters](https://api.otto.de/portal/guidelines/r100024)
- Added description and reference to the content negotiation mechanism
of HTTP that is being leveraged with profile versioning.
- Updated all example containing profile and link relations URLs
- The rule "Must use kebab-case profile URIs" has been removed because
it is a duplication of https://api.otto.de/portal/guidelines/r000023
- The rule "SHOULD NOT use resource versioned path" has been removed
because it is a duplication of [SHOULD NOT use URI
versioning](https://api.otto.de/portal/guidelines/r000026)

Changelog:
<!--changelogbody-->
- Renamed rule "MUST use absolute URLs for custom link relation types"
to "MUST use absolute URIs for custom link relation types". The
requirement for a `https://api.otto.de/portal` prefix has been removed,
and it is no longer required to provide resolvable link relation URIs.
(R100037)
- The "Profile" section previously being found in "Hypermedia" has been
merged into the section "Versioning of incompatible changes" because
profiles are used for versioning of hypermedia and non-hypermedia APIs.
- Renamed rule "MUST use profiles for Public REST APIs" to "MUST use
profiles for versioning" to better indicate the purpose of profiles. The
content describing the concept of a profile in depth has been moved from
the section "Profile" to this rule. Added further description and
examples to highlight better how to leverage profiles. (R000065)
- The rule "MUST use resolvable profile URLs" has been moved to
"Compatibility"->"Versioning of incompatible changes" and renamed to
"MUST use absolute URIs for profiles". The requirement for a
`https://api.otto.de/portal` prefix for profile URIs has been removed,
and it is no longer required to provide resolvable profile URIs.
(R100066)
- Removed the rule "MAY use type and profile attributes in hyperlinks"
because the cardinality of the profile link may change and profile URIs
are no longer resolvable.
- Removed the rule "Must use kebab-case profile URIs" as it is a
duplicate of "MUST use kebab-case for URIs". (R000023)
- Removed the rule "SHOULD NOT use resource versioned path" as it is a
duplicate of "SHOULD NOT use URI versioning". (R000026)

---------

Co-authored-by: Thorsten Hake <mail@thorsten-hake.com>
Co-authored-by: serviceintegration-bulldozer[bot] <60896867+serviceintegration-bulldozer[bot]@users.noreply.github.com>
Co-authored-by: Birgit Bader <birgit.bader@otto.de>